Added concept of "Sequencer"
The idea is to flip the responsibilities for more complex animation sequence. Instead of having the Entity stack validate the sequence every frame, it offload the sequence to the "Sequencer" which run continuously unless interrupted at specific "cut" points.

The Entity can then interrogate the sequencer to insert actions at certain points in parallel.

If the sequencer doesn't hear from the Entity for more than 2 frames it gives up and abort.