Remade shopping
Fixed PutInFridge? So close but still bugging out a bit
Interesting issue with trying to go buy more foodstuff when there's already some in the fridge. The ActionPlan is considered finished if there's already foodstuff in the fridge. But the fridge still allow the ActionPlan but reduces the satisfaction reward. Should it NOT allow the plan or should I find a way to allow more than one foodstuff in the fridge?